requirements.txt
streamlit
google-generativeai
groq
google-genai

